--- packages/modeling/src/operations/extrusions/slice/toTriangles.d.ts
@@ -1,7 +1,0 @@
-import Poly3 from '../../../geometries/poly3/type'
-
-import Slice from './type'
-
-export default toTriangles
-
-declare function toTriangles(slice: Slice): Array<Poly3>
